fix(settings): avoid routing policy stack conflicts
