+++ app/db/models.py
@@ -263,12 +263,6 @@ class DashboardSettings(Base):
         server_default=false(),
         nullable=False,
     )
-    sticky_reallocation_budget_threshold_pct: Mapped[float] = mapped_column(
-        Float,
-        default=95.0,
-        server_default=text("95.0"),
-        nullable=False,
-    )
     sticky_reallocation_primary_budget_threshold_pct: Mapped[float] = mapped_column(
         Float,
         default=95.0,
@@ -279,6 +273,12 @@ class DashboardSettings(Base):
         Float,
         default=100.0,
         server_default=text("100.0"),
+        nullable=False,
+    )
+    sticky_reallocation_budget_threshold_pct: Mapped[float] = mapped_column(
+        Float,
+        default=95.0,
+        server_default=text("95.0"),
         nullable=False,
     )
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now(), nullable=False)
