+++ tests/integration/test_settings_api.py
@@ -81,55 +81,3 @@ async def test_settings_api_get_and_update(async_client):
     assert payload["totpRequiredOnLogin"] is False
     assert payload["totpConfigured"] is False
     assert payload["apiKeyAuthEnabled"] is True
-
-
-@pytest.mark.asyncio
-async def test_settings_legacy_sticky_threshold_updates_primary_threshold(async_client):
-    response = await async_client.put(
-        "/api/settings",
-        json={
-            "stickyThreadsEnabled": True,
-            "preferEarlierResetAccounts": True,
-            "stickyReallocationBudgetThresholdPct": 88.0,
-        },
-    )
-
-    assert response.status_code == 200
-    updated = response.json()
-    assert updated["stickyReallocationBudgetThresholdPct"] == 88.0
-    assert updated["stickyReallocationPrimaryBudgetThresholdPct"] == 88.0
-    assert updated["stickyReallocationSecondaryBudgetThresholdPct"] == 100.0
-
-
-@pytest.mark.asyncio
-async def test_settings_primary_sticky_threshold_updates_legacy_threshold(async_client):
-    response = await async_client.put(
-        "/api/settings",
-        json={
-            "stickyThreadsEnabled": True,
-            "preferEarlierResetAccounts": True,
-            "stickyReallocationPrimaryBudgetThresholdPct": 87.0,
-        },
-    )
-
-    assert response.status_code == 200
-    updated = response.json()
-    assert updated["stickyReallocationBudgetThresholdPct"] == 87.0
-    assert updated["stickyReallocationPrimaryBudgetThresholdPct"] == 87.0
-    assert updated["stickyReallocationSecondaryBudgetThresholdPct"] == 100.0
-
-
-@pytest.mark.asyncio
-async def test_settings_full_put_honors_changed_legacy_sticky_threshold(async_client):
-    response = await async_client.get("/api/settings")
-    assert response.status_code == 200
-    payload = response.json()
-    payload["stickyReallocationBudgetThresholdPct"] = 86.0
-
-    response = await async_client.put("/api/settings", json=payload)
-
-    assert response.status_code == 200
-    updated = response.json()
-    assert updated["stickyReallocationBudgetThresholdPct"] == 86.0
-    assert updated["stickyReallocationPrimaryBudgetThresholdPct"] == 86.0
-    assert updated["stickyReallocationSecondaryBudgetThresholdPct"] == 100.0
